Improve handling of fractional and integer parts
This changes the behavior of the Calculator when the
query includes numbers that have only a fractional part
or integer part.

Assume the decimal place is represented by a period.

`I` and `F` refer to the integer and fractional parts
respectively.

When the query is only a number, and that number is of
the form `I.`, the Goodie will reject the input. This is because
there is very little extra information that can be provided.

When the query is of the form `.F`, the fraction representation
of the decimal number will be returned - e.g, an input of
`.625` will result in an answer of `5/8`.

This also tightens-up the triggering of the Calculator based on
numbers in the query.